Fix clangd warnings in include/ and src/ files (#14)
* Fix clangd warnings in render_manager of type casting, err printing, and RenderText input argument structure

* Fix clang warnings in scene by casting types explicitly

* Fix clangd warnings in physics manager by updating type casting and function signature of UpdateEnemyState

* Fix clangd warnings in collision manager by changing types of num_X vars and casting loop var i to int

* Fix clangd warnings in entity manager with type casting and UpdateObservation redundant argument removal

* Fix clangd warnings in action manager by removing redundant arg from GetActionSize

* Remove unused arg in CalculateTotalReward and add std::move to avoid copying twice in AddTerms

* Cast std::numbers::pi to float

* Add std::move to name of BaseProjSpell to avoid double copy and add missing switch case to SpellUpgradeType for count

* Fix tests given the new function signatures for calculating total reward and getting action size

* Fix adjacents parameters having similar types in ItemStatUpgrade

* Change ItemValueRange into more generic ValueRange to be used in spell upgrades and fix corresponding tests

* Change function signature of SetGrid to use SDL_Rect

* Add explicit type casting to float and remove unnecessary std::move

* Update function sig of obs size getter due to unused input arg

* Update formatting in progression manager using linter

* Update raycaster to use Ray struct to avoid having input args with similar types

* Update observation manager tests with new obs size getter signature

* Add new RayHistoryIndex struct to avoid having similarly typed args next to eachother

* Update rewards based on new RayHistoryIndex struct

* Update type casting to avoid errors

* Update type casting in utils of the get avg frame time func

* Comment out unused args of main

* Fix the type casting in game.cpp and comment out unused arg in the SignalHandler

* Update function signatures in the pybindings and change obs/act buffer to be const ref

+++ include/scene.h
@@ -66,14 +66,14 @@ struct Scene {
 
     // Add slight variation to each enemy to make it more interesting.
     for (int i = 0; i < kNumEnemies; ++i) {
-      enemy.movement_speed[i] += GenerateRandomInt(1, 100);
+      enemy.movement_speed[i] += static_cast<float>(GenerateRandomInt(1, 100));
       int random_width = GenerateRandomInt(1, 50);
       int random_height = GenerateRandomInt(1, 50);
       enemy.sprite_size[i].width += random_width;
       enemy.sprite_size[i].height += random_height;
       // TODO: Figure out a more maintainable solution to collider + sprite size randomization.
-      enemy.collider[i].offset.x += 0.5f * random_width;
-      enemy.collider[i].offset.y += 0.5f * random_height;
+      enemy.collider[i].offset.x += 0.5f * static_cast<float>(random_width);
+      enemy.collider[i].offset.y += 0.5f * static_cast<float>(random_height);
       enemy.collider[i].size.width += random_width;
       enemy.collider[i].size.height += random_height;
     };
